+++ translation_rag/__main__.py
@@ -1,0 +1,5 @@
+"""CLI entry point for Translation RAG."""
+from .cli import main
+
+if __name__ == "__main__":
+    main()
